+++ src/vscode/filesystemProvider/fabric/FabricFSItemType.ts
@@ -5,33 +5,29 @@ import { FabricFSCacheItem } from './FabricFSCacheItem';
 import { FabricFSWorkspace } from './FabricFSWorkspace';
 import { FabricApiService } from '../../../fabric/FabricApiService';
 import { FabricApiItemType } from '../../../fabric/_types';
+import { FABRIC_FS_ITEM_TYPES } from './_types';
 
 export class FabricFSItemType extends FabricFSCacheItem {
-	workspace: FabricFSWorkspace;
-	itemType: FabricApiItemType;
-
 	constructor(uri: FabricFSUri) {
 		super(uri);
-
-		this._stats = {
-			type: vscode.FileType.Directory,
-			ctime: undefined,
-			mtime: undefined,
-			size: undefined
-		};
 	}
 
-	get workspaceId(): string {
-		return this.workspace.id;
+	get itemType(): FabricApiItemType {
+		return this.FabricUri.itemType;
 	}
 
 	public async loadStatsFromApi<T>(): Promise<void> {
-		this._stats = {
-			type: vscode.FileType.Directory,
-			ctime: undefined,
-			mtime: undefined,
-			size: undefined
-		};
+		if (FABRIC_FS_ITEM_TYPES.includes(this.itemType)) {
+			this._stats = {
+				type: vscode.FileType.Directory,
+				ctime: undefined,
+				mtime: undefined,
+				size: undefined
+			};
+		}
+		else {
+			this._stats = undefined;
+		}
 	}
 
 	public async loadChildrenFromApi<T>(): Promise<void> {